Remove bundle plugin and use package.json version
