+++ src/plugins/navigation.ts
@@ -3,7 +3,7 @@ import { definePlugin } from '../plugin.js';
 
 export const navigationPlugin = definePlugin({
   name: 'navigation',
-  version: '0.1.0',
+
   description: 'React Navigation / Expo Router state inspection',
 
   async setup(ctx) {
