+++ src/version.ts
@@ -1,0 +1,2 @@
+import pkg from '../package.json';
+export const version: string = pkg.version;
